requirements.txt
aiohttp>=3.9.0
websockets>=12.0
pyyaml>=6.0
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
aiogram>=3.4.0
pydantic>=2.0
